# Use strings.Replacer for a large efficiency gain