--- a/format_ja_jp.go
+++ b/format_ja_jp.go
@@ -77,7 +77,7 @@ func parseFuncJaCommon(locale Locale) internalParseFunc {
 			value = strings.Replace(value, k, v, -1)
 		}
 
-		value = commonFormatFunc(value, layout,
+		value = commonParseFunc(value, layout,
 			knownDaysShortReverse[locale], knownDaysLongReverse[locale],
 			knownMonthsShortReverse[locale], knownMonthsLongReverse[locale], knownPeriods[locale])