--- a/format_zh_cn.go
+++ b/format_zh_cn.go
@@ -68,7 +68,7 @@ func parseFuncZhCommon(locale Locale) internalParseFunc {
 			value = strings.Replace(value, k, v, -1)
 		}
 
-		return commonFormatFunc(value, layout,
+		return commonParseFunc(value, layout,
 			knownDaysShortReverse[locale], knownDaysLongReverse[locale],
 			knownMonthsShortReverse[locale], knownMonthsLongReverse[locale], knownPeriods[locale])
 	}