By using strings.Replacer, we can drastically reduce the number of allocations and increase performance in Format for non-genitive languages. For these languages, formatting is more-or-less just a find and replace operation, so this change has a huge impact.

Benchstat shows the following results:

```
name      old time/op    new time/op    delta
Format-8    2.03ms ±12%    0.39ms ± 4%  -80.94%  (p=0.000 n=9+9)

name      old alloc/op   new alloc/op   delta
Format-8     593kB ± 0%      93kB ± 0%  -84.35%  (p=0.000 n=9+10)

name      old allocs/op  new allocs/op  delta
Format-8     20.0k ± 0%      2.7k ± 0%  -86.44%  (p=0.002 n=8+10)
```

That's a pretty big increase in performance. It undersells the gain too; the benchmark includes a wide variety of languages. 

While the genitive languages and parsing functions can still be improved by a good margin, the common formatting case is mostly optimized. Though it's possible to take it a step further. 

For example, strings.Replacer works on strings and thus copies the string once per case; if we had a bytes.Replacer equivalent and used []byte internally, that could be faster. We could then add an `AppendFormat` function to reduce allocations to zero for those who want it.