@ztstewart 2019-07-14
Not sure why it's showing the README / linter changes in the PR as those have already landed on my branch. Should be no-ops (and git diff doesn't show them).

Tests pass:

```
zstewart@LAPTOP-J9J9VKM0:ztstewart/monday ‹zts/efficiency*›$ go test -race ./...
ok      github.com/ztstewart/monday     (cached)
````