@ztstewart 2019-07-14
I guess the one question I have is if the tests cover cases where this logic might break things. Maybe it's possible for some odd cases, e.g. duplicate format specifiers or something. I may add more test cases