@ztstewart 2019-07-14
After some additional testing with new test cases, a few uncommon scenarios will behave strangely. If a user specifies both a short month and long month in the format string or both a long day and short day in the format string, this change may cause the format to render incorrectly. It's a pretty unusual case, in my experience, but one that can indeed happen.

I have a potential fix that I will push to this PR at some point after more vetting. Bonus: it appears to make formatting a bit faster than what's in here.